+++ app/models/ductwork/job.rb
@@ -89,7 +89,7 @@ module Ductwork
         job_klass: klass
       )
       args = JSON.parse(input_args)["args"]
-      instance = Object.const_get(klass).new(*args)
+      instance = Object.const_get(klass).new(args)
       run = execution.create_run!(
         started_at: Time.current
       )