+++ spec/ductwork/job_spec.rb
@@ -86,7 +86,7 @@ RSpec.describe Ductwork::Job do
       expect(job.klass).to eq("MyFirstStep")
       expect(job.started_at).to be_within(1.second).of(Time.current)
       expect(job.completed_at).to be_nil
-      expect(job.input_args).to eq(JSON.dump({ args: args }))
+      expect(job.input_args).to eq(JSON.dump({ args: }))
       expect(job.output_payload).to be_nil
       expect(job.step).to eq(step)
     end