+++ spec/ductwork/pipeline_spec.rb
@@ -250,14 +250,6 @@ RSpec.describe Ductwork::Pipeline do
       expect(Ductwork::Job).to have_received(:enqueue).with(anything, 1)
     end
 
-    it "correctly passes multiple arguments to the job" do
-      allow(Ductwork::Job).to receive(:enqueue)
-
-      klass.trigger(1, 2)
-
-      expect(Ductwork::Job).to have_received(:enqueue).with(anything, 1, 2)
-    end
-
     it "correctly passes an array argument to the job" do
       allow(Ductwork::Job).to receive(:enqueue)